Add Support for "mediation" (#24)
Thanks. That's great. Mediation was indeed something I wanted to add since a while. I'll merge this for now, but it's likely I'll rework it in the future: check if mediation is available on the browser first, improve the docs, add it to the demo... there's always kind of many other things to do besides just adding the flag. I'll also like to experiment with the different values myself with several browsers / platforms to get a better feel for it.